+++ src/style/style.css
@@ -21,6 +21,7 @@ body {
   display: flex;
   flex: 1;
   max-width: 300px;
+  min-width: 300px;
   height: 100%;
   flex-direction: column;
   overflow-x: scroll;
@@ -65,7 +66,6 @@ ul {
 }
 
 .documentTreeButtonContainer {
-  /* opacity: 0; */
   display: flex;
   justify-content: center;
   align-items: center;
@@ -84,6 +84,7 @@ ul {
   font-size: 20px;
   transition: all 0.3s;
 }
+
 .addDocumentButton:hover,
 .deleteDocumentButton:hover {
   cursor: pointer;
@@ -136,6 +137,7 @@ ul {
 #editor {
   flex: 2;
   background-color: #f7ffe5;
+  min-width: 300px;
 }
 
 .editorContainer {
