+++ 수정할거.md
@@ -2,8 +2,10 @@
 2. ui수정
 3. documnetId에대한 validation?
 4. 컴포넌트의 불변성?
-5. 컴포넌트 자체를 넘겨주지 않으면서 state를 설정하는 법?
-6. DocumentTree 렌더링 innerHTML로 바꿔서 렌더링 방식을 통일해주자
-7. 무한 input
-8. link클릭 li전체로 바꾸기
-9. DocumentTree 내부 요소인 DocumentTreeBranch 생성하기 -> 7/8 완료
+5. DocumentTree 렌더링 innerHTML로 바꿔서 렌더링 방식을 통일해주자
+6. 무한 input
+7. link클릭 li전체로 바꾸기 -> 7/8일 완료
+8. DocumentTree 내부 요소인 DocumentTreeBranch 생성하기 -> 7/8 완료
+9. route를 좀더 세심하게 만들어봐 친구야..
+10. Component의 props에 대한 validation -> 7/8일 완료
+11. render역할 줄이기
